requirements.txt
absl-py==2.2.2
accelerate==1.0.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
airportsdata==20250224
annotated-types==0.7.0
anyio==4.6.2.post1
astor==0.8.1
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
blake3==1.0.4
bs4==0.0.2
cachetools==5.5.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.8
cloudpickle==3.1.1
compressed-tensors==0.9.3
contourpy==1.3.0
cupy-cuda12x==13.4.1
cycler==0.12.1
datasets==3.0.2
Deprecated==1.2.18
depyf==0.18.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
einops==0.8.0
email_validator==2.2.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
fastapi==0.115.12
fastapi-cli==0.0.7
fastrlock==0.8.3
filelock==3.16.1
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.9.0
gguf==0.14.0
googleapis-common-protos==1.70.0
grpcio==1.71.0
h11==0.14.0
hf-xet==1.0.3
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.30.2
idna==3.10
imgkit==1.2.3
importlib_metadata==8.0.0
interegular==0.3.3
Jinja2==3.1.6
jiter==0.6.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
lark==1.2.2
llguidance==0.7.14
llvmlite==0.44.0
lm-format-enforcer==0.10.11
loguru==0.7.2
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
mistral_common==1.5.4
modelscope==1.21.1
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.19.0
multidict==6.1.0
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.4
nltk==3.9.1
numba==0.61.2
numpy==2.1.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai==1.52.2
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
openpyxl-image-loader==1.0.5
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-exporter-otlp-proto-http==1.26.0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-semantic-conventions-ai==0.4.3
outcome==1.3.0.post0
outlines==0.1.11
outlines_core==0.1.26
packaging==24.1
pandas==2.2.3
partial-json-parser==0.2.1.1.post5
peft==0.13.2
pillow==11.0.0
pip==24.2
prometheus_client==0.21.1
prometheus-fastapi-instrumentator==7.1.0
propcache==0.2.0
protobuf==4.25.6
psutil==6.1.0
py-cpuinfo==9.0.0
pyarrow==18.0.0
pycountry==24.6.1
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.19.1
PyMetis==2023.1.1
pyparsing==3.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.4.0
ray==2.43.0
referencing==0.36.2
regex==2024.9.11
requests==2.32.3
rich==14.0.0
rich-toolkit==0.14.1
rouge_score==0.1.2
rpds-py==0.24.0
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
selenium==4.27.1
sentence-transformers==4.1.0
sentencepiece==0.2.0
setuptools==75.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
starlette==0.46.2
sympy==1.13.1
tabulate==0.9.0
threadpoolctl==3.5.0
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.66.6
transformers==4.51.3
trio==0.28.0
trio-websocket==0.11.1
triton==3.2.0
typer==0.15.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.34.1
uvloop==0.21.0
watchfiles==1.0.5
websocket-client==1.8.0
websockets==15.0.1
wheel==0.44.0
wrapt==1.17.2
wsproto==1.2.0
xformers==0.0.29.post2
xgrammar==0.1.18
xxhash==3.5.0
yarl==1.17.0
zipp==3.21.0